// prototype/src/app/models/parametre.model.ts
export default class Parametre {
	key?: string | null;
	  nom?: string;
	  description?: string;
	  color?: string;
	  logo?: any;
	  url?: string;
	  contact?: string;
	  adresse?: any;
	  type?: any;
	  tiktok?: any;
	  facebook?: any;
	  instagram?: any;
	  tweeter?: any;
	  linkedin?: any;
	  youtube?: any;
}